fix: Client bugs + moved server code to src folder.
