+++ server/src/reducers/store.ts
@@ -1,5 +1,5 @@
-import storeRedux from '../../shared/src/reducers/store'
-import { storeUpdateSettings } from '../../shared/src/actions/settingsActions'
+import storeRedux from '../../../shared/src/reducers/store'
+import { storeUpdateSettings } from '../../../shared/src/actions/settingsActions'
 import { loadSettings } from '../utils/SettingsStorage'
 
 storeRedux.dispatch(storeUpdateSettings(loadSettings(storeRedux.getState())))
